delete: 기존 도메인 삭제 및 패키지 이동
