+++ domain/src/main/kotlin/omock/lib/ark/ArkBlackWinRule.kt
@@ -1,10 +1,10 @@
-package lib.ark
+package omock.lib.ark
 
-object ArkOverLineRule : ArkRule() {
+object ArkBlackWinRule : ArkRule() {
     override fun validate(
         board: List<List<Int>>,
         position: Pair<Int, Int>,
-    ): Boolean = directions.all { direction -> !checkWhiteWin(board, position, direction) }
+    ): Boolean = directions.map { direction -> checkWhiteWin(board, position, direction) }.contains(true)
 
     private fun checkWhiteWin(
         board: List<List<Int>>,
@@ -14,8 +14,9 @@ object ArkOverLineRule : ArkRule() {
         val oppositeDirection = direction.let { (dx, dy) -> Pair(-dx, -dy) }
         val (stone1, blink1) = search(board, position, oppositeDirection)
         val (stone2, blink2) = search(board, position, direction)
+
         return when {
-            blink1 + blink2 == 0 && stone1 + stone2 > 4 -> true
+            blink1 + blink2 == 0 && stone1 + stone2 == 4 -> true
             else -> false
         }
     }
