+++ domain/src/main/kotlin/omock/lib/ark/ArkOverLineRule.kt
@@ -1,12 +1,12 @@
-package lib.ark
+package omock.lib.ark
 
-object ArkWhiteWinRule : ArkRule(WHITE_STONE, BLACK_STONE) {
+object ArkOverLineRule : ArkRule() {
     override fun validate(
         board: List<List<Int>>,
         position: Pair<Int, Int>,
-    ): Boolean = directions.map { direction -> checkBlackWin(board, position, direction) }.contains(true)
+    ): Boolean = directions.all { direction -> !checkWhiteWin(board, position, direction) }
 
-    private fun checkBlackWin(
+    private fun checkWhiteWin(
         board: List<List<Int>>,
         position: Pair<Int, Int>,
         direction: Pair<Int, Int>,
@@ -14,9 +14,8 @@ object ArkWhiteWinRule : ArkRule(WHITE_STONE, BLACK_STONE) {
         val oppositeDirection = direction.let { (dx, dy) -> Pair(-dx, -dy) }
         val (stone1, blink1) = search(board, position, oppositeDirection)
         val (stone2, blink2) = search(board, position, direction)
-
         return when {
-            blink1 + blink2 == 0 && stone1 + stone2 >= 4 -> true
+            blink1 + blink2 == 0 && stone1 + stone2 > 4 -> true
             else -> false
         }
     }
