+++ domain/src/main/kotlin/omock/lib/constant/LibraryConstant.kt
@@ -1,4 +1,4 @@
-package lib.ark.constant
+package omock.lib.constant
 
 object LibraryConstant {
     const val BOARD_SIZE = 15
