+++ domain/src/main/kotlin/omock/model/Stone.kt
@@ -1,0 +1,6 @@
+package omock.model
+
+enum class Stone {
+    BLACK,
+    WHITE,
+}
