+++ domain/src/main/kotlin/omock/model/Vector.kt
@@ -1,4 +1,4 @@
-package omok.model
+package omock.model
 
 data class Vector(val x: Int, val y: Int) {
     operator fun unaryMinus(): Vector = Vector(-x, -y)
