--- domain/src/main/kotlin/omok/model/OmokStone.kt
@@ -1,3 +1,0 @@
-package omok.model
-
-data class OmokStone(val position: Position, val color: StoneColor)
