--- domain/src/main/kotlin/omok/model/StoneColor.kt
@@ -1,7 +1,0 @@
-package omok.model
-
-enum class StoneColor {
-    BLACK,
-    WHITE,
-    NONE,
-}
