--- domain/src/main/kotlin/omok/model/game/OmokGameResult.kt
@@ -1,7 +1,0 @@
-package omok.model.game
-
-sealed interface OmokGameResult
-
-data object Placed : OmokGameResult
-
-data object InvalidGameRule : OmokGameResult
