--- domain/src/test/kotlin/omok/fixtures/PointFixtures.kt
@@ -1,8 +1,0 @@
-package omok.fixtures
-
-import omok.model.Position
-
-fun createPosition(
-    x: Int,
-    y: Int,
-) = Position.of(x, y)
